Move files into source dir
